requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
datasets==2.17.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
evaluate==0.4.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter-client==8.6.0
jupyter-core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.0
progressbar2==4.3.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
responses==0.18.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
triton==2.2.0
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
